requirements2.txt
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
comm==0.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.6
imageio==2.5.0
imgaug==0.2.9
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-segmentation==0.3.0
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.1.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.8
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
PyWavelets==1.1.1
pywin32==305
pywinpty==1.1.6
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.5.4
Send2Trash==1.8.2
Shapely==1.8.5.post1
six==1.15.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.36.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.6
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
